return a 400 when the Accept header is invalid
